Comments and code cleanup

+++ src/problems/1-50/9/problem9.cpp
@@ -5,11 +5,7 @@
 #include "problem9.hpp"
 
 namespace problems {
-    /**
-     * Specifies values, solves problem and outputs solution and calculation time.
-     * @return The exit code
-     */
-    ulong problem9::solve(uint sum) {
+    uint problem9::solve(uint sum) {
         for (uint a = 1; a <= sum / 3; a++)
             for (uint b = a + 1; b < sum / 2; b++) {
                 uint c = sum - a - b;
